requirements.txt
django~=3.1.4
social-auth-app-django~=4.0.0
python-jose~=3.2.0
python-dotenv~=0.15.0
django-sslserver